requirements.txt
# Web Server
fastapi~=0.127.0
uvicorn[standard]
python-multipart

# Gemini & Data Handling
google-genai>=0.3.0
pydantic>=2.0.0
Pillow~=12.0.0

# Utilities
python-dotenv~=1.2.1